+++ packages/create-app/templates/lit/src/index.js
@@ -1,4 +1,5 @@
-import { LitElement, customElement, css, html } from 'lit-element';
+import { LitElement, css, html } from 'lit';
+import { customElement } from 'lit/decorators.js';
 
 @customElement('my-element')
 export class MyElement extends LitElement {