+++ packages/plugin-litcss/CHANGELOG.md
@@ -1,2 +1,5 @@
+## 0.21.0
+- Uses `css` function from `lit` package
+
 ## 0.18.2
 - First release