+++ packages/plugin-malinajs/CHANGELOG.md
@@ -1,5 +1,8 @@
+## 0.21.0
+- Target Malina.js `v0.6.43`
+
 ## 0.19.2
-- Target Malina.js `0.6.10`
+- Target Malina.js `v0.6.10`
 
 ## 0.19.0
 - Target Malina.js `v0.6.8`