+++ packages/plugin-prefresh/CHANGELOG.md
@@ -1,2 +1,5 @@
+## 0.21.0
+- Update prefresh dependency
+
 ## 0.20.0
 - First release