+++ packages/plugin-typescript/CHANGELOG.md
@@ -1,2 +1,5 @@
+## 0.21.0
+- Target TypeScript `v4.4.3`
+
 ## 0.19.0
 - First release